Adjust max values of X, Y and pressure
Set them to the actual max values that the tablet can report.

This gets the tablet working better!
The pointer now moves on the whole screen.
When the pen is out of range, the pointer is moved to the upper left corner.
This is caused by the inRange boolean always being true.